Add optional parameter for root certificate (chain)
The vSphere API provides the optional 'root_cert' parameter when issuing
a PUT request to '/api/vcenter/certificate-management/vcenter/tls'.
This parameter allows supplying a Root CA certificate (chain) for the
updated machine certificate.
This functionality was not represented in vCenterTls before.
This was resolved by adding the parameter 'RootCert' and associated
logic to the 'Set-vCenterCertificate' function.